git: Raise exception on ref checkout from dirty repo
This raises an exception on trying to checkout a ref from a
dirty repo in the git_handler module. The parent patch
https://review.openstack.org/#/c/582652/ currently forcibly
cleans the repo but this is undesirable as it may have local
user changes that need to be resolved first.

The safest path is for Pegleg to immediately raise an exception
on any tracked/untracked files that are detected using the
GitPython API.

Unit tests are added for both untracked/tracked file cases.

Change-Id: I2241bc981dca1999508c3c7e95948fe47a5ddebf