git: Use currently checked out ref when none supplied
This patch set adds logic to examine the currently checked out
reference in a repository if ref=None is passed to git.git_handler.

This makes it convenient to simply reuse the revision that is
currently checked out without having to re-supply it. Note that
ref is still a required parameter so providing None will have
to be done explicitly.

Defaults for ref='master' have been changed to ref=None.
Turns out the GitPython library has logic to use the
repo's Git config's fetch refspec by default [0],
which means that by default 'master' will be checked out
for many repositories.

Also adds a ``is_repository`` function to ``git`` module to
more robustly check whether a directory is a repo.

Unit tests and doc strings are updated as well.

[0] https://github.com/gitpython-developers/GitPython/blob/a8591a094a768d73e6efb5a698f74d354c989291/git/remote.py#L779

Change-Id: Ib1736eb20a51996cf03fa7f6187f078b39595267