fix: Always check for exit_code 0 in CLI tests
This patch set fixes a recent issue that was introducing
following a refactor Pegleg table output. The related
CLI unit tests were changed to stop checking for an exit_code
of 0 due to improper mocking causing an exit_code of -1
to be returned instead.

This patch set corrects the issue by refactoring the code to
allow for exit_code 0 to always be checked.

Further refactoring was done in places to reduce code
duplication via helpers.

Change-Id: Ib9cf7c71b02d74de152b7e34911c867d2251a18a