refactor: Allow site_by_params to take in list of fields
This patch set simplifies some code related to site_by_params
by allowing it to take an iterable argument called *fields
which specifies exactly which parameters to include from
the site-definition.yaml. This means that no hard-coding
is required to manually filter the params into the exact
parameters required by follow-up function calls. This is
done for better code maintenance.

Change-Id: Ief6483dfbf3759204106330284e8e9b824b5567e