Fix --save-location error in decrypt command
When the --save-location flag is used it attempts to keep the name
of the file being decrypted, and saving it to a desired location.
Previously this would cause
TypeError: join() argument must be str or bytes, not 'tuple'

This is resolved by grabbing the correct value from the os.path.split
command.

Change-Id: I9344bd07f5f8d03b50ac9fc004b79fefb024bbd6