Update dockerfiles to include ssh client
It was discovered that some base images when used as an override to
the specified default images do not include the openssh package.
This is particularly problematic if a user wishes to use ssh access
when specifying their repositories in the site's site-definition.yaml

Without the openssh package the following error occurs:
ERROR pegleg.engine.util.git:normalize_repo_path [nnn] The
repo_path=ssh://user@domain:port/site-repo is not a valid Git repo

Adding the openssh package does not impact the current base images as
they already include it, but has the added benefit of ensuring that
a non-default base image will still work with Pegleg.

Change-Id: I154c3db5071a373ad16cb0a0c4c6103b6ea8ac4e