Pegleg Shipyard Auth Token instead of user/pass
Recently Shipyard was updated in [0] to allow a user to provide an
auth token manually, as an environment variable, instead of having
Shipyard obtain a token from Keystone using passed in auth variables
including a username and password.

This patch expands on that functionality to allow a user to either
set an OS_AUTH_TOKEN environment variable, or pass it during the
upload command with the --os-auth-token flag. If it is present, only
the token is used and the other auth variables are ignored.

[0] https://review.opendev.org/#/c/666402/

Change-Id: Iaf2a109022e8f5d496851ff43fcfa8198b5411c9