Allow Pegleg to overwrite existing files
Pegleg currently raises an exception when a filepath exists before
writing to it in the case of yaml.dump actions.  This is not always
desired behavior, and with the addition of the flags passed into the
files methods for write/append is no longer necessary.

Removing these so that overwrite options can take place for example
when rotating passphrases or secrets.

Change-Id: I2200cc681a797d91a90f6237662fd1ee77954906