fix: Fix Pegleg pep8 gate
This commit updates the Pegleg pep8 gate to use Ubuntu 18.04 (bionic) as
the base Ubuntu image for the job. Upstream CI has a community goal to
use Ubuntu 20.04 (focal) which defaults to python3.8. By moving to
python3.8, some of the pip packages need to be updated but until pip
conflicts are resolved, it is difficult to do so. Aside from that, the
functional gates are continuing to use Ubuntu 18.04.

Additionally, update the deb-docker path.

Change-Id: I599d8b8973a893a2112ff1219e10c19534c2c962